I've added a program for warming up to a maximal effort single and also the option of a 15 lb technique bar.
Warming up for a single can often times be a bit confusing for novice
lifters, but maxes provide a great way of testing growth over a training
period. The template is based off of a T-Nation article and is meant to
be a suggestion template for those who haven't attempted many max effort
lifts before.

The inclusion of a 15 lb technique bar might not seem like much, but a
few pounds can be pretty significant when the lifter is using working
weights under 100 lbs.

I have tested my changes, and I haven't noticed any side effects.